requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-groq
langchain-community
langchain-huggingface
langchain-chroma
langchain-text-splitters
chromadb
pypdf
python-dotenv
fastmcp
sentence-transformers
pydantic
python-multipart